ViewMap und AddGeoTag zusammengefasst

+++ test/GeoTagging/src/com/testrunns/geotagging/GeoTag.java
@@ -59,7 +59,7 @@ public class GeoTag {
 		return longitude;
 	}
 
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 
@@ -67,7 +67,7 @@ public class GeoTag {
 		return latitude;
 	}
 
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
